+++ CHANGELOG.md
@@ -5,7 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://github.com/fivetran/terraform-provider-fivetran/compare/v1.1.17...HEAD)
+## [Unreleased](https://github.com/fivetran/terraform-provider-fivetran/compare/v1.1.18...HEAD)
+
+## [1.1.18](https://github.com/fivetran/terraform-provider-fivetran/compare/v1.1.17...v1.1.18)
+
+## Updated
+- Schema updated resource `resource_connector_schema_config`:
+    - Field `schema` deprecated dues to issue with Set type performance
+    - Field `schemas` of Map type added instead.
 
 ## [1.1.17](https://github.com/fivetran/terraform-provider-fivetran/compare/v1.1.16...v1.1.17)